feat(api): Add authorization token

+++ api/src/module/facture/facture.controller.ts
@@ -1,23 +1,27 @@
-import {Body, Controller, Delete, Get, Param, Post, Put} from "@nestjs/common";
+import {Body, Controller, Delete, Get, Param, Post, Put, Headers} from "@nestjs/common";
 import {ApiBadRequestResponse, ApiCreatedResponse, ApiOkResponse, ApiOperation, ApiTags} from "@nestjs/swagger";
 import {FactureModel} from "./facture.model";
 import {FactureService} from "./facture.service";
+import {TokenValidation} from "../../validation/token/token.validation";
 
 @Controller({path: 'facture'})
 @ApiTags('facture')
 export class FactureController {
 
     private readonly factureService: FactureService;
+    private readonly tokenValidation: TokenValidation;
 
     constructor() {
         this.factureService = new FactureService();
+        this.tokenValidation = new TokenValidation();
     }
 
     @Get()
     @ApiOperation({summary: 'Get multiple factures'})
     @ApiOkResponse({description: 'List of factures'})
     @ApiBadRequestResponse({description: 'Request param is not valid'})
-    getFactures() {
+    async getFactures(@Headers('authorization') token: string) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.factureService.getFactures();
     }
 
@@ -25,7 +29,8 @@ export class FactureController {
     @ApiOperation({summary: 'Create facture'})
     @ApiCreatedResponse({description: 'Facture created'})
     @ApiBadRequestResponse({description: 'Request body is not valid'})
-    createFacture(@Body() body: FactureModel) {
+    async createFacture(@Headers('authorization') token: string, @Body() body: FactureModel) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.factureService.createFacture(body);
     }
 
@@ -33,7 +38,8 @@ export class FactureController {
     @ApiOperation({summary: 'Update facture'})
     @ApiOkResponse({description: 'Facture updated'})
     @ApiBadRequestResponse({description: 'Request body is not valid'})
-    updateFacture(@Param('id') id: number, @Body() body: FactureModel) {
+    async updateFacture(@Headers('authorization') token: string, @Param('id') id: number, @Body() body: FactureModel) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.factureService.updateFacture(id, body);
     }
 
@@ -41,7 +47,8 @@ export class FactureController {
     @ApiOperation({summary: 'Delete facture'})
     @ApiOkResponse({description: 'Facture deleted'})
     @ApiBadRequestResponse({description: 'Request param is not valid'})
-    deleteFacture(@Param('id') id: number) {
+    async deleteFacture(@Headers('authorization') token: string, @Param('id') id: number) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.factureService.deleteFacture(id);
     }
 }
