+++ api/src/module/location/location.controller.ts
@@ -1,22 +1,26 @@
-import {Body, Controller, Delete, Get, Param, Post, Put} from "@nestjs/common";
+import {Body, Controller, Delete, Get, Headers, Param, Post, Put} from "@nestjs/common";
 import {ApiBadRequestResponse, ApiCreatedResponse, ApiOkResponse, ApiOperation, ApiTags} from "@nestjs/swagger";
 import {LocationService} from "./location.service";
 import {LocationModel} from "./location.model";
+import {TokenValidation} from "../../validation/token/token.validation";
 
 @Controller({path: 'location'})
 @ApiTags('Location')
 export class LocationController {
     private locationService: LocationService;
+    private tokenValidation: TokenValidation;
 
     constructor() {
         this.locationService = new LocationService();
+        this.tokenValidation = new TokenValidation();
     }
 
     @Get()
     @ApiOperation({summary: 'Get multiple locations'})
     @ApiOkResponse({description: 'List of locations'})
     @ApiBadRequestResponse({description: 'Request param is not valid'})
-    getLocations() {
+    async getLocations(@Headers('authorization') token: string) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.locationService.getLocations();
     }
 
@@ -24,7 +28,8 @@ export class LocationController {
     @ApiOperation({summary: 'Create location'})
     @ApiCreatedResponse({description: 'Location created'})
     @ApiBadRequestResponse({description: 'Request body is not valid'})
-    createLocation(@Body() body: LocationModel) {
+    async createLocation(@Headers('authorization') token: string, @Body() body: LocationModel) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.locationService.createLocation(body);
     }
 
@@ -32,7 +37,8 @@ export class LocationController {
     @ApiOperation({summary: 'Update location'})
     @ApiOkResponse({description: 'Location updated'})
     @ApiBadRequestResponse({description: 'Request body is not valid'})
-    updateLocation(@Param('id') id: number, @Body() body: LocationModel) {
+    async updateLocation(@Headers('authorization') token: string, @Param('id') id: number, @Body() body: LocationModel) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.locationService.updateLocation(id, body);
     }
 
@@ -40,7 +46,8 @@ export class LocationController {
     @ApiOperation({summary: 'Delete location'})
     @ApiOkResponse({description: 'Location deleted'})
     @ApiBadRequestResponse({description: 'Request param is not valid'})
-    deleteLocation(@Param('id') id: number) {
+    async deleteLocation(@Headers('authorization') token: string, @Param('id') id: number) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.locationService.deleteLocation(id);
     }
 }
