+++ api/src/module/service/service.controller.ts
@@ -1,23 +1,27 @@
-import {Body, Controller, Delete, Get, Param, Post, Put} from "@nestjs/common";
+import {Body, Controller, Delete, Get, Headers, Param, Post, Put} from "@nestjs/common";
 import {ApiBadRequestResponse, ApiOkResponse, ApiOperation, ApiTags} from "@nestjs/swagger";
 import {ServiceService} from "./service.service";
 import {ServiceModel} from "./service.model";
+import {TokenValidation} from "../../validation/token/token.validation";
 
 @Controller('service')
 @ApiTags('Service')
 export class ServiceController {
 
     private serviceService: ServiceService;
+    private tokenValidation: TokenValidation;
 
     constructor() {
         this.serviceService = new ServiceService();
+        this.tokenValidation = new TokenValidation();
     }
 
     @Get()
     @ApiOperation({summary: 'Get multiple services'})
     @ApiOkResponse({description: 'List of services'})
     @ApiBadRequestResponse({description: 'Request param is not valid'})
-    async getServices() {
+    async getServices(@Headers('authorization') token: string) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.serviceService.getServices();
     }
 
@@ -25,7 +29,8 @@ export class ServiceController {
     @ApiOperation({summary: 'Create service'})
     @ApiOkResponse({description: 'Service created'})
     @ApiBadRequestResponse({description: 'Request body is not valid'})
-    async createService(@Body() body: ServiceModel) {
+    async createService(@Headers('authorization') token: string, @Body() body: ServiceModel) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.serviceService.createService(body);
     }
 
@@ -33,7 +38,8 @@ export class ServiceController {
     @ApiOperation({summary: 'Update service'})
     @ApiOkResponse({description: 'Service updated'})
     @ApiBadRequestResponse({description: 'Request body is not valid'})
-    async updateService(@Param('id') id: number, @Body() body: ServiceModel) {
+    async updateService(@Headers('authorization') token: string, @Param('id') id: number, @Body() body: ServiceModel) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.serviceService.updateService(id, body);
     }
 
@@ -41,7 +47,8 @@ export class ServiceController {
     @ApiOperation({summary: 'Delete service'})
     @ApiOkResponse({description: 'Service deleted'})
     @ApiBadRequestResponse({description: 'Request param is not valid'})
-    async deleteService(@Param('id') id: number) {
+    async deleteService(@Headers('authorization') token: string, @Param('id') id: number) {
+        await this.tokenValidation.validateAdminToken(token);
         return this.serviceService.deleteService(id);
     }
